requirements.txt
PySide6>=6.5.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
requests>=2.31.0
lxml>=4.9.0 
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
geopy>=2.3.0
folium>=0.14.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0 